+++ moe/bandit/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""Bandit directory containing multi-armed bandit implementation in python.
+
+Contains:
+
+    * :mod:`moe.bandit.epsilon.Epsilon`
+    * :mod:`moe.bandit.epsilon_greedy.EpsilonGreedy`
+
+"""