+++ moe/views/rest/gp_ei.py
@@ -13,7 +13,8 @@ from pyramid.view import view_config
 from moe.optimal_learning.python.cpp_wrappers.expected_improvement import ExpectedImprovement
 from moe.optimal_learning.python.timing import timing_context
 from moe.views.constant import GP_EI_ROUTE_NAME, GP_EI_PRETTY_ROUTE_NAME
-from moe.views.gp_pretty_view import GpPrettyView, PRETTY_RENDERER
+from moe.views.gp_pretty_view import GpPrettyView
+from moe.views.pretty_view import PRETTY_RENDERER
 from moe.views.schemas.rest.gp_ei import GpEiRequest, GpEiResponse
 from moe.views.utils import _make_gp_from_params
 
@@ -60,7 +61,7 @@ class GpEiView(GpPrettyView):
            :input: :class:`moe.views.schemas.rest.GpEiRequest`
            :output: :class:`moe.views.schemas.rest.GpEiResponse`
 
-           :status 201: returns a response
+           :status 200: returns a response
            :status 500: server error
 
         """